+++ primefaces-selenium/primefaces-selenium-core/src/main/java/org/primefaces/selenium/internal/junit/PageInjectionExtension.java
@@ -23,10 +23,6 @@
  */
 package org.primefaces.selenium.internal.junit;
 
-import java.lang.reflect.Field;
-
-import javax.inject.Inject;
-
 import org.junit.jupiter.api.extension.*;
 import org.openqa.selenium.WebDriver;
 import org.primefaces.selenium.AbstractPrimePage;
@@ -34,7 +30,32 @@ import org.primefaces.selenium.PrimeSelenium;
 import org.primefaces.selenium.spi.PrimePageFactory;
 import org.primefaces.selenium.spi.WebDriverProvider;
 
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Field;
+import java.util.ArrayList;
+import java.util.List;
+
 public class PageInjectionExtension implements ParameterResolver, TestInstancePostProcessor {
+
+    private final List<Class<? extends Annotation>> injectMarkerAnnotations;
+
+    public PageInjectionExtension() {
+        injectMarkerAnnotations = new ArrayList<>();
+
+        try {
+            injectMarkerAnnotations.add((Class<? extends Annotation>) Class.forName("javax.inject.Inject"));
+        }
+        catch (ClassNotFoundException e) {
+            // Ignore
+        }
+
+        try {
+            injectMarkerAnnotations.add((Class<? extends Annotation>) Class.forName("jakarta.inject.Inject"));
+        }
+        catch (ClassNotFoundException e) {
+            // Ignore
+        }
+    }
 
     @Override
     public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext)
@@ -61,7 +82,7 @@ public class PageInjectionExtension implements ParameterResolver, TestInstancePo
         for (Field field : testInstance.getClass().getDeclaredFields()) {
             field.setAccessible(true);
 
-            if (field.getAnnotation(Inject.class) != null
+            if ((injectMarkerAnnotations.stream().anyMatch(it -> field.getAnnotation(it) != null))
                         && AbstractPrimePage.class.isAssignableFrom(field.getType())
                         && field.get(testInstance) == null) {
 
@@ -73,5 +94,4 @@ public class PageInjectionExtension implements ParameterResolver, TestInstancePo
             }
         }
     }
-
 }